+++ schemars_derive/src/attr/mod.rs
@@ -7,9 +7,8 @@ pub use schemars_to_serde::process_serde_attrs;
 use proc_macro2::{Group, Span, TokenStream, TokenTree};
 use syn::parse::{self, Parse};
 
-pub fn get_with_from_attrs(field: &syn::Field) -> Option<syn::Result<syn::Type>> {
-    field
-        .attrs
+pub fn get_with_from_attrs(attrs: &[syn::Attribute]) -> Option<syn::Result<syn::Type>> {
+    attrs
         .iter()
         .filter(|at| match at.path.get_ident() {
             // FIXME this is relying on order of attributes (schemars before serde) from schemars_to_serde.rs